Refactor editor hooks
