+++ app/neo/client/useNeoClient.test.tsx
@@ -2,7 +2,8 @@ import { useLocation, useNavigate } from '@remix-run/react';
 import { act, renderHook } from '@testing-library/react';
 import { Callback, NeoClient } from '~/data/neo-protocol';
 import { Process } from '~/data/process-api';
-import { useCreateEditor, useEditors } from '../editors/useEditors';
+import { useCreateEditor } from '../editors/useCreateEditor';
+import { useEditors } from '../editors/useEditors';
 import { AnimationFollowMode } from '../settings/useSettings';
 import { NeoClientProviderContext, useNeoClient } from './useNeoClient';
