+++ app/neo/client/useNeoClient.tsx
@@ -3,7 +3,8 @@ import { useLocation, useNavigate } from '@remix-run/react';
 import { createContext, useContext } from 'react';
 import { NeoClientJsonRpc } from '~/data/neo-jsonrpc';
 import { NeoClient } from '~/data/neo-protocol';
-import { useCreateEditor, useEditors } from '~/neo/editors/useEditors';
+import { useEditors } from '~/neo/editors/useEditors';
+import { useCreateEditor } from '../editors/useCreateEditor';
 import { useWebSocket } from '../editors/useWebSocket';
 import { AnimationFollowMode } from '../settings/useSettings';
