+++ app/routes/$ws.dataclasses.$app.$pmv.$/route.tsx
@@ -1,7 +1,7 @@
 import editorStylesHref from '@axonivy/variable-editor/lib/style.css?url';
 import type { MetaFunction } from '@remix-run/node';
 import { LinksFunction } from '@remix-run/node';
-import { useRestoreEditor } from '~/neo/editors/useEditors';
+import { useRestoreEditor } from '~/neo/editors/useRestoreEditor';
 
 export const links: LinksFunction = () => [{ rel: 'stylesheet', href: editorStylesHref }];
