+++ app/routes/$ws.forms._index/route.tsx
@@ -4,7 +4,8 @@ import { FormIdentifier, useCreateForm, useDeleteForm, useForms } from '~/data/f
 import { ProjectIdentifier } from '~/data/project-api';
 import { ArtifactCard, cardLinks, NewArtifactCard } from '~/neo/artifact/ArtifactCard';
 import { useNewArtifact } from '~/neo/artifact/useNewArtifact';
-import { Editor, useCreateEditor, useEditors } from '~/neo/editors/useEditors';
+import { useCreateEditor } from '~/neo/editors/useCreateEditor';
+import { Editor, useEditors } from '~/neo/editors/useEditors';
 import { Overview } from '~/neo/Overview';
 import PreviewSVG from './form-preview.svg?react';
