+++ app/routes/$ws.processes.$app.$pmv.$/route.tsx
@@ -1,5 +1,5 @@
 import type { MetaFunction } from '@remix-run/node';
-import { useRestoreEditor } from '~/neo/editors/useEditors';
+import { useRestoreEditor } from '~/neo/editors/useRestoreEditor';
 
 export const meta: MetaFunction = () => {
   return [{ title: 'Axon Ivy Process' }, { name: 'description', content: 'Axon Ivy Process' }];
